Revert "Update dev-wifi from my fork"
