+++ src/graphics/Screen.h
@@ -17,7 +17,6 @@
 #include "concurrency/PeriodicTask.h"
 #include "power.h"
 #include <string>
-#include <WiFi.h>
 
 namespace graphics
 {
@@ -47,7 +46,6 @@ class DebugInfo
     /// Renders the debug screen.
     void drawFrame(OLEDDisplay *display, OLEDDisplayUiState *state, int16_t x, int16_t y);
     void drawFrameSettings(OLEDDisplay *display, OLEDDisplayUiState *state, int16_t x, int16_t y);
-    void drawFrameWiFi(OLEDDisplay *display, OLEDDisplayUiState *state, int16_t x, int16_t y);
 
 
     std::string channelName;
@@ -221,8 +219,6 @@ class Screen : public concurrency::PeriodicTask
     static void drawDebugInfoTrampoline(OLEDDisplay *display, OLEDDisplayUiState *state, int16_t x, int16_t y);
 
     static void drawDebugInfoSettingsTrampoline(OLEDDisplay *display, OLEDDisplayUiState *state, int16_t x, int16_t y);
-
-    static void drawDebugInfoWiFiTrampoline(OLEDDisplay *display, OLEDDisplayUiState *state, int16_t x, int16_t y);
 
     /// Queue of commands to execute in doTask.
     TypedQueue<ScreenCmd> cmdQueue;
