+++ src/main.cpp
@@ -41,8 +41,6 @@
 #include "timing.h"
 #include <OneButton.h>
 #include <Wire.h>
-#include "meshwifi/meshwifi.h"
-#include "meshwifi/meshhttp.h"
 // #include <driver/rtc_io.h>
 
 #ifndef NO_ESP32
@@ -327,9 +325,6 @@ void setup()
     }
 #endif
 
-    // Initialize Wifi
-    initWifi();
-
     if (!rIf)
         recordCriticalError(ErrNoRadio);
     else
@@ -420,9 +415,6 @@ void loop()
     // FIXME - until button press handling is done by interrupt (see polling above) we can't sleep very long at all or buttons
     // feel slow
     msecstosleep = 10;
-
-    // TODO: This should go into a thread handled by FreeRTOS.
-    handleWebResponse();
 
     delay(msecstosleep);
 }
