+++ src/main.h
@@ -10,8 +10,6 @@ extern bool ssd1306_found;
 extern bool isCharging;
 extern bool isUSBPowered;
 
-
-
 // Global Screen singleton.
 extern graphics::Screen screen;
 //extern Observable<meshtastic::PowerStatus> newPowerStatus; //TODO: move this to main-esp32.cpp somehow or a helper class
